# public/openapi.yaml
openapi: 3.0.3
info:
  title: HPKV Cloud API
  description: API endpoints for interacting with an HPKV store, mirroring hpkvfs logic.
  version: 0.1.0
servers:
  - url: /api # Relative path to the API routes within the Next.js app
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: hpkv-api-key
      description: Your HPKV API Key.
    ApiUrlAuth:
      type: apiKey
      in: header
      name: hpkv-api-url
      description: The base URL of your HPKV API endpoint (e.g., https://api-eu-1.hpkv.io).
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Description of the error.
security:
  - ApiKeyAuth: []
  - ApiUrlAuth: []
paths:
...
  /write:
    post:
      summary: Write file content
      description: Writes data to a file at a specified offset. Creates the file if it doesn't exist. Expects base64 encoded data in the request body.
      parameters:
        - name: path
          in: query
          required: true
          description: The absolute path of the file to write to.
          schema:
            type: string
        - name: offset
          in: query
          required: true
          description: The byte offset at which to start writing.
          schema:
            type: integer
            format: int64
            minimum: 0
      requestBody:
        description: Base64 encoded data to write.
        required: true
        content:
          text/plain:
            schema:
              type: string
              format: base64
      responses:
        '200':
          description: Successfully wrote data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  bytesWritten:
                    type: integer
                    format: int32
        '400':
          description: Bad Request (e.g., missing parameters, invalid offset, invalid base64, path is directory).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error (e.g., failed to write chunk, failed to update metadata).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...